[PATCH] USB: RNDIS (and CDC) filter flag handling
This should fix the problem David Meggy found, where RNDIS was setting
the OID_GEN_CURRENT_PACKET_FILTER state incorrectly.  It's the same
issue Andrew Morton noticed a while back, for that matter, but with
more than just a "now compiles on 64 bit" fix.

Basically the code needs to interpret 32 bits provided in the request
from the (Windows) host, rather than 8 bits of other memory that's got
some irrelevant value.

The fix is just to save the 32 bits.  I did the same thing with the
CDC Ethernet filter, which should eventually be used the same way:  to
limit what packets get sent to the host.  Also defined a couple more
of the CDC requests.