[PATCH] dynamic addition of virtual disks on PPC64 iSeries
From: Stephen Rothwell <sfr@canb.auug.org.au>

This patch allows us to dynamically add virtual disks to an iSeries partition.
 It works like this: after you have created the virtual disk file on OS/400
and attached it to the Linux partition, you need to write to
/sys/bus/vio/drivers/viodasd/probe (it doesn't matter what you write).  This
will do the probe.  It calls add_disk() for each new disk, so we get hotplug
events as a side effect.

This was the nicest way I could think of doing this as the interface to the
hypervisor is polled ...