[PATCH] USB: Make usb_choose_configuration() a separate subroutine
This patch creates a separate subroutine to encapsulate the work required
in choosing an initial configuration for a device.  In my opinion that
reflects a policy decision that should be kept apart from all the other
work involved in creating a new device structure.  The existing code in
usb_new_device() is simply moved into usb_choose_configuration().